% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sf.R
\name{design_sf}
\alias{design_sf}
\title{Make sf from design}
\usage{
design_sf(design, polygon, dim1_var = RUN, dim2_var = RANGE)
}
\arguments{
\item{design}{A data.frame}

\item{polygon}{An sf polygon}

\item{dim1_var}{The (unquoted) name of the variable describing the first dimension of the design}

\item{dim2_var}{The (unquoted) name of the variable describing the second dimension of the design}
}
\description{
Make sf from design
}
