% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{furthest}
\alias{furthest}
\title{furthest}
\usage{
furthest(polygon, axis = 1, reverse = FALSE)
}
\arguments{
\item{polygon}{An sf polygon}

\item{axis}{1 is north/south, 2 is east/west}

\item{reverse}{will return the furthest east or south unless this is TRUE}
}
\value{
An integer defining which coordinate is furthest on the axis
}
\description{
furthest
}
\examples{

}
