% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_points}
\alias{make_points}
\title{Make points}
\usage{
make_points(dim1, dim2, dim1x, dim2x, dim1y, dim2y, refx, refy,
  moddim1 = 0, moddim2 = 0)
}
\arguments{
\item{dim1}{}

\item{dim2}{}

\item{dim1x}{}

\item{dim2x}{}

\item{dim1y}{}

\item{dim2y}{}

\item{refx}{}

\item{refy}{}

\item{moddim1}{}

\item{moddim2}{}
}
\value{
A length two vector
}
\description{
Make points
}
