% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_polygon}
\alias{make_polygon}
\title{Make a polygon}
\usage{
make_polygon(dim1, dim2, dim1x, dim2x, dim1y, dim2y, refx, refy,
  coords = NULL)
}
\arguments{
\item{dim1}{}

\item{dim2}{}

\item{dim1x}{}

\item{dim2x}{}

\item{dim1y}{}

\item{dim2y}{}

\item{refx}{}

\item{refy}{}

\item{coords}{}
}
\value{

}
\description{
Make a polygon
}
